+++ CHANGELOG.md
@@ -70,6 +70,15 @@
         let = null
         ```
 
+    * It is no longer a compile-time error to assign to an import when not bundling:
+
+        ```js
+        import {foo} from 'path'
+        foo = null
+        ```
+
+        Instead the behavior will be left up to the host environment at run-time, which should cause a run-time error. However, this will still be treated as a compile-time error when bundling because the scope-hoisting optimization that happens during bundling means the host may no longer cause run-time errors.
+
 ## 0.6.30
 
 * Fix optional call of `super` property ([#362](https://github.com/evanw/esbuild/issues/362))